refactor: don't use too-light or too-dark colors

+++ scripts/deploy.sh
@@ -2,9 +2,10 @@
 # prep work: created a git worktree in ./build, added a dummy commit, pushed it
 # to origin
 set -euo pipefail
+rm -f /tmp/.git
 if [ -f build/.git ]; then cp build/.git /tmp/.git; fi
 pnpm build # <- this nukes the .git
-cp /tmp/.git ./build/
+if [ -f /tmp/.git ]; then cp /tmp/.git ./build/; fi
 cd ./build
 git add .
 git commit -m "chore: deploy"
